x/badge/types/message_create_series_test.go
package types

import (
	"testing"

	"github.com/concrnt/concord/testutil/sample"
	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
	"github.com/stretchr/testify/require"
)

func TestMsgCreateSeries_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgCreateSeries
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgCreateSeries{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgCreateSeries{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}
